feat: پشتیبانی از مقدار اعشاری در فیلد حق شیفت (PCT)
## مشکل
فیلد «حق شیفت» در حکم کارمندی فقط اعداد صحیح را می‌پذیرفت.
کاربر نمی‌توانست مقدار ۷.۵٪ (مثلاً) وارد کند چون:
- ستون AMOUNT در PAY2_DECREE_LINE از نوع BIGINT بود
- DTO مدل از نوع long بود
- SP_PAY2_CALC_RUN متغیر @ITEM_AMOUNT را BIGINT تعریف کرده بود
- UI هم اجازه ورود اعشار نمی‌داد

## تغییرات

### پایگاه داده
- Migration 008: تبدیل PAY2_DECREE_LINE.AMOUNT از BIGINT به DECIMAL(18,2)
  (idempotent — فقط در صورتی اجرا می‌شود که ستون هنوز BIGINT باشد)
- ScriptSqly.cs: اصلاح schema برای نصب‌های تازه (DECIMAL(18,2))
- ScriptSqly.cs: اصلاح SP_PAY2_CALC_RUN — متغیر @ITEM_AMOUNT به DECIMAL(18,2)
  (این SP هر بار startup از نو ساخته می‌شود — تغییر اینجا الزامی بود)

### پشته سرور
- Program.cs: افزودن 008_DecreeLineAmountDecimal.sql به لیست migration خودکار

### مدل مشترک
- Pay2DecreeLineDto.AMOUNT: از long به decimal

### رابط کاربری (Blazor)
- InputMode=Percentage برای آیتم‌های shift — اجازه ورود اعشار
- MaxLength=3 + MaxDecimalPlaces=2 → EffectiveMaxLength=6 (حداکثر "100.50")
- UpdateShiftHint: long.TryParse → decimal.TryParse با InvariantCulture
- SubmitSave: long.TryParse → decimal.TryParse با InvariantCulture
- GetMonthlyBaseSalary: نوع برگشتی long? → decimal?
- EditLine: AMOUNT.ToString با InvariantCulture (نمایش صحیح اعشار)
- Placeholder و Tooltip به‌روز شدند

### فایل‌های مرجع (PAY2_DDL_v6.sql, PAY2_Procedures_v6.sql)
- به‌روزرسانی برای هماهنگی با schema واقعی

+++ Server/Info/PAY2_DDL_v6.sql
@@ -576,7 +576,7 @@ CREATE TABLE [dbo].[PAY2_DECREE_LINE]
 (
     [DEC_ID]   INT      NOT NULL,
     [ITEM_ID]  INT      NOT NULL,
-    [AMOUNT]   BIGINT   NOT NULL CONSTRAINT DF_DL_AMT DEFAULT(0),
+    [AMOUNT]   DECIMAL(18,2) NOT NULL CONSTRAINT DF_DL_AMT DEFAULT(0),
     [INS_OV]   BIT      NULL,                                                -- NULL=از PAY2_ITEM_DEF
     [TAX_OV]   BIT      NULL,
     [BASIS_OV] TINYINT  NULL,
